% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_uuid}
\alias{as_uuid}
\title{Turn any vector of strings into a UUID}
\usage{
as_uuid(s)
}
\arguments{
\item{s}{vector of strings}
}
\value{
UUID
}
\description{
Turn any vector of strings into a UUID
}
