% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-file-builder.R
\name{bsub_prefix}
\alias{bsub_prefix}
\title{Generate a bsub-compatible prefix.}
\usage{
bsub_prefix(c = comment)
}
\value{
prefix for BSUB option.
}
\description{
Generate a bsub-compatible prefix.
}
