% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-install.R
\name{build_cmdstan}
\alias{build_cmdstan}
\title{Build `stanc` in the cmdstan directory.}
\usage{
build_cmdstan(config_file)
}
\arguments{
\item{config_file}{.yaml file listing how to install CmdStan.}
}
\value{
config object of install.
}
\description{
Build `stanc` in the cmdstan directory.
}
