% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{build_registry}
\alias{build_registry}
\title{Build a registry}
\usage{
build_registry(root, name = "registry", target = NULL)
}
\arguments{
\item{root}{root of the tree of Stannis runs (where to find hash dirs)}

\item{name}{name of the registry file (w/o extension)}

\item{target}{optional directory where to place the registry}
}
\value{
path to the new registry
}
\description{
Build a registry
}
