% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-r-hat.R
\name{calculate_parameter_psrf}
\alias{calculate_parameter_psrf}
\title{Calculate Potential Scale Reduction Factor (PSRF)}
\usage{
calculate_parameter_psrf(chains)
}
\arguments{
\item{set}{the chains to operate on}
}
\value{
array of PSRF values for each entry of the parameter.
}
\description{
Works on a single parameter from a set of chains.
}
