% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{calculate_plist_reduction}
\alias{calculate_plist_reduction}
\title{Calculate Potential Scale Reduction Factor (PSRF) or 
other per-parameter-scalar summary.}
\usage{
calculate_plist_reduction(x, f = split_rhat_rfun, drop = c("stepsize__",
  "treedepth__", "n_leapfrog__", "divergent__", "iteration"))
}
\arguments{
\item{set}{the chains to operate on.}
}
\value{
list of PSRF arrays (one per parameter)
}
\description{
Works on an entire set of parameters with specific list
structure
}
