% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-r-hat.R
\name{calculate_set_psrf}
\alias{calculate_set_psrf}
\title{Calculate Potential Scale Reduction Factor (PSRF)}
\usage{
calculate_set_psrf(set)
}
\arguments{
\item{set}{the chains to operate on.}
}
\value{
list of PSRF arrays (one per parameter)
}
\description{
Works on an entire set of parameters.
}
