% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{check_key}
\alias{check_key}
\title{Check that a key exists in the registry}
\usage{
check_key(registry, key)
}
\arguments{
\item{registry}{a registry object}

\item{hash}{the key (hash) to check}
}
\value{
logical (TRUE/FALSE)
}
\description{
Check that a key exists in the registry
}
