% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-compile.R
\name{compile_model}
\alias{compile_model}
\title{Return binary path to a compiled model, don't recompile if
the stashed model is the same as the to-be-compiled model
according to a hash (just return the path).}
\usage{
compile_model(args)
}
\arguments{
\item{args}{an argument tree}
}
\value{
path to the binary of the compiled model.
}
\description{
Return binary path to a compiled model, don't recompile if
the stashed model is the same as the to-be-compiled model
according to a hash (just return the path).
}
