% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{construct_cmdline}
\alias{construct_cmdline}
\title{Create a flat string of CmdStan arguments based on a 
sub-tree of an arg-tree.}
\usage{
construct_cmdline(...)
}
\arguments{
\item{...}{arguments to run CmdStan with.}
}
\value{
string of arguments to CmdStan.
}
\description{
Create a flat string of CmdStan arguments based on a 
sub-tree of an arg-tree.
}
