% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{copy_failed}
\alias{copy_failed}
\title{Throw error when copying fails.}
\usage{
copy_failed(from, to)
}
\arguments{
\item{from}{source path}

\item{to}{destination path}
}
\description{
Throw error when copying fails.
}
