% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{covariate_block-class}
\alias{covariate_block}
\alias{covariate_block-class}
\title{A reference class for constructing the (simple!) model matrix for a
covariate based on a model formula (default is additive of all
covariates.}
\description{
A reference class for constructing the (simple!) model matrix for a
covariate based on a model formula (default is additive of all
covariates.
}
\section{Fields}{

\describe{
\item{\code{formula}}{The R-style formula used to describe the linear covariate
relationship.  Typically lacking an intercept in models made of
multiple blocks.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(name, covariate = list(), validation = NULL, drops = NULL)}}{Initialize the base class with a name and validation function.}

\item{\code{make_block()}}{For the base class this is a placeholder which throws an error.
In all other subclasses this function arranges the covariates into the
class-specific model matrix.  The name of the model}
}}

