% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{create_array_set}
\alias{create_array_set}
\title{Extract array set from a Stan file set.}
\usage{
create_array_set(set)
}
\arguments{
\item{set}{object created by `read_stan_set`}
}
\value{
list of arrays, one per chain in set
}
\description{
Extract array set from a Stan file set.
}
