% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R
\name{create_hash}
\alias{create_hash}
\title{create a pseudo-'secret', salted with project_id}
\usage{
create_hash(args)
}
\arguments{
\item{args}{arg-tree object (list).}
}
\value{
hash
}
\description{
This is a hash of the:
  1) project id
  2) model file contents
  3) model binary contents
  4) data file contents (if any)
  5) init file contents (if any)
  6) arbitrary hash salt
  7) prepared but not finalized arg tree
}
