% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dim,stan_commander-method}
\alias{dim,stan_commander-method}
\title{Accessor to get dimensions of parameter for the currently chosen id/type.}
\usage{
\S4method{dim}{stan_commander}(x)
}
\arguments{
\item{x}{stan_commander object.}
}
\description{
Accessor to get dimensions of parameter for the currently chosen id/type.
}

