% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{extension}
\alias{extension}
\title{Retrieve just file extension from a string}
\usage{
extension(s)
}
\arguments{
\item{s}{string}

\item{split}{what to split the extension with}
}
\value{
extension
}
\description{
Retrieve just file extension from a string
}
