% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{extract_array}
\alias{extract_array}
\title{Extract a single parameter array from a Stan file set.}
\usage{
extract_array(set, parameter)
}
\arguments{
\item{set}{object created by `read_stan_set`}
}
\value{
list of arrays, one per chain in set
}
\description{
Extract a single parameter array from a Stan file set.
}
