% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cn-helpers.R
\name{extract_parameter_names}
\alias{extract_parameter_names}
\title{Extract unique parameter names from a set of Stan-format
column names.}
\usage{
extract_parameter_names(x)
}
\arguments{
\item{x}{Stan-format column names.}
}
\value{
unique parameter names.
}
\description{
Extract unique parameter names from a set of Stan-format
column names.
}
