% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg-helpers.R
\name{finalize_args}
\alias{finalize_args}
\title{Finalize an argument tree object by merging components
that are not fully specified.}
\usage{
finalize_args(args)
}
\arguments{
\item{args}{arg-tree object (list).}
}
\value{
same tree, ready to pass to 'construct_cmdline'
        and 'run_model_cmd'
}
\description{
Finalize an argument tree object by merging components
that are not fully specified.
}
