% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{find_}
\alias{find_}
\title{Find entry at a path in a list}
\usage{
find_(x, path = NULL)
}
\arguments{
\item{x}{the list}

\item{path}{(in-list) to the desired element}
}
\value{
location of the name
}
\description{
Find entry at a path in a list
}
