% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-compile.R
\name{find_cmdstan}
\alias{find_cmdstan}
\title{Find CmdStan installation}
\usage{
find_cmdstan(args)
}
\arguments{
\item{args}{argument list, with possible hints..}
}
\value{
CmdStan directory
}
\description{
Find CmdStan installation
}
