% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice-model.R
\name{find_file}
\alias{find_file}
\title{Return first candidate file in search directories}
\usage{
find_file(search, name)
}
\arguments{
\item{search}{directories to look in (comma-separated string).}

\item{name}{file name to search for.}
}
\value{
first candidate matching name.
}
\description{
Return first candidate file in search directories
}
