% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{find_run_files}
\alias{find_run_files}
\title{Find the components of a single run under a set of paths}
\usage{
find_run_files(root, control = "finalized.yaml", output = "output.csv",
  diagnostics = "diagnostics.csv")
}
\arguments{
\item{root}{root path to search under}

\item{control}{name of the control files}

\item{output}{name of the output files}

\item{name}{of the diagnostics files}
}
\value{
data frame indexing the file sets
}
\description{
Find the components of a single run under a set of paths
}
