% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg-helpers.R
\name{flatten_args}
\alias{flatten_args}
\title{expand chains, inits, or both}
\usage{
flatten_args(args)
}
\arguments{
\item{args}{agr-tree object (list).}
}
\value{
list of arg-trees, separated s.t. each arg-tree
        is for one chain and one init file.
}
\description{
expand chains, inits, or both
}
