% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-helpers.R
\name{flatten_array}
\alias{flatten_array}
\title{Replace array dimensionality by explicit long-format data labels.}
\usage{
flatten_array(A, ...)
}
\arguments{
\item{A}{array to process}

\item{...}{further args passed to "data.frame"}
}
\value{
data frame with all array data in long format.
}
\description{
Replace array dimensionality by explicit long-format data labels.
}
