% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-file-builder.R
\name{generate_option}
\alias{generate_option}
\title{Generate an option selection}
\usage{
generate_option(prefix, name, value)
}
\arguments{
\item{prefix}{the function to generate the right prefix for
the platform}

\item{name}{the name of the option to set}

\item{value}{the value to set the option to}
}
\value{
string encoding the option.
}
\description{
Generate an option selection
}
