% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-file-builder.R
\name{generate_options}
\alias{generate_options}
\title{generate a set of options based on arguments}
\usage{
generate_options(ostream, args, prefix = pbs_prefix)
}
\arguments{
\item{ostream}{where to write the string}

\item{args}{list of arguments to generate to stream}

\item{prefix}{function to generate prefix with}
}
\value{
output stream path
}
\description{
generate a set of options based on arguments
}
