% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-df-helpers.R
\name{generate_parameter_array}
\alias{generate_parameter_array}
\title{Create an array from a Stan-style data frame for a 
given parameter.}
\usage{
generate_parameter_array(data)
}
\arguments{
\item{data}{data frame of the set of columns for a given
parameter.}
}
\value{
array of correct dimension for the given parameter
        indexed as in the Stan model.
}
\description{
Create an array from a Stan-style data frame for a 
given parameter.
}
