% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-df-helpers.R
\name{generate_parameter_arrays}
\alias{generate_parameter_arrays}
\title{Generate one properly sized array for each Stan parameter
in a given Stan-format data frame.}
\usage{
generate_parameter_arrays(data)
}
\arguments{
\item{data}{stan .csv file loaded as data.frame}
}
\value{
list of arrays (element per named parameter).
}
\description{
Generate one properly sized array for each Stan parameter
in a given Stan-format data frame.
}
