% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cn-helpers.R
\name{generate_stan_column_name}
\alias{generate_stan_column_name}
\title{Convert parameter name and index into Stan-style column
name.}
\usage{
generate_stan_column_name(parameter, ...)
}
\arguments{
\item{parameter}{name of parameter}

\item{...}{indices to use (i=1:10, j=3:7).}
}
\value{
character vector of column names.
}
\description{
Convert parameter name and index into Stan-style column
name.
}
