% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_adaptation_used}
\alias{get_adaptation_used}
\title{Whether adaptation is engaged}
\usage{
get_adaptation_used(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
TRUE if adaptation was used.
}
\description{
Whether adaptation is engaged
}
