% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_algorithm_name}
\alias{get_algorithm_name}
\title{Algorithm name from processed list.}
\usage{
get_algorithm_name(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
algorithm name
}
\description{
Algorithm name from processed list.
}
