% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-tree-helpers.R
\name{get_binary_dir}
\alias{get_binary_dir}
\title{return binary dir for this project from an argument 
tree ('args' object).  If the binary_dir is missing, 
create an R-session-specific tempdir.}
\usage{
get_binary_dir(args)
}
\arguments{
\item{args}{arg-tree object (list).}
}
\value{
project binary dir
}
\description{
return binary dir for this project from an argument 
tree ('args' object).  If the binary_dir is missing, 
create an R-session-specific tempdir.
}
