% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-install.R
\name{get_cmdstan}
\alias{get_cmdstan}
\title{Get CmdStan from github.com/stan-dev/cmdstan}
\usage{
get_cmdstan(config_file)
}
\arguments{
\item{config_file}{.yaml file listing where to get
CmdStan from.}
}
\value{
config object
}
\description{
Get CmdStan from github.com/stan-dev/cmdstan
}
