% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-helpers.R
\name{get_column_names}
\alias{get_column_names}
\title{Get the line of a Stan .csv file with the column names in it, starts
with "lp__".}
\usage{
get_column_names(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get 
past the recorded meta-data.}
}
\value{
vector of all column names, with whitespace trimmed.
}
\description{
Get the line of a Stan .csv file with the column names in it, starts
with "lp__".
}
