% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-helpers.R
\name{get_comment_lines}
\alias{get_comment_lines}
\title{Get the lines of a Stan .csv file with the metadata}
\usage{
get_comment_lines(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
only the lines starting with a hash.
}
\description{
Get the lines of a Stan .csv file with the metadata
}
