% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-helpers.R
\name{get_control_lines}
\alias{get_control_lines}
\title{Get the lines of a Stan .csv file with the control parameters.}
\usage{
get_control_lines(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
only the lines with the A=B format.
}
\description{
Get the lines of a Stan .csv file with the control parameters.
}
