% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_delta}
\alias{get_delta}
\title{Delta adaptation parameter}
\usage{
get_delta(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
Delta adaptation parameter
}
