% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_diagnostic_file}
\alias{get_diagnostic_file}
\title{diagnostic file path}
\usage{
get_diagnostic_file(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
path
}
\description{
diagnostic file path
}
