% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Read the parameter dimensions of the binary rewrite}
\usage{
get_dimensions(root)
}
\arguments{
\item{root}{where the rewrite is rooted.}
}
\value{
named list of parameter dimensions
}
\description{
Read the parameter dimensions of the binary rewrite
}
