% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_engine_name}
\alias{get_engine_name}
\title{Engine name from processed list.}
\usage{
get_engine_name(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
engine name
}
\description{
Engine name from processed list.
}
