% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{get_expectations}
\alias{get_expectations}
\title{Return the expected outputs from a job:}
\usage{
get_expectations(job, logger)
}
\arguments{
\item{job}{a job}

\item{logger, }{a logger}
}
\value{
character vector of expected object names
}
\description{
Return the expected outputs from a job:
}
