% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{get_expected_files}
\alias{get_expected_files}
\title{Returns the expected output files from a job:}
\usage{
get_expected_files(job, logger)
}
\arguments{
\item{job}{a job}

\item{logger, }{a logger}
}
\value{
character vector of expected object names
}
\description{
Returns the expected output files from a job:
}
