% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_gamma}
\alias{get_gamma}
\title{Gamma adaptation parameter}
\usage{
get_gamma(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
Gamma adaptation parameter
}
