% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-tree-helpers.R
\name{get_id}
\alias{get_id}
\title{return project id from an argument tree ('args' object).}
\usage{
get_id(args)
}
\arguments{
\item{args}{arg-tree object (list).}
}
\value{
project id.
}
\description{
return project id from an argument tree ('args' object).
}
