% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-helpers.R
\name{get_imm_diagonal}
\alias{get_imm_diagonal}
\title{Get the line of a Stan .csv file with the inverse mass matrix
diagonal in it.}
\usage{
get_imm_diagonal(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
vector of values of diagonal.
}
\description{
Get the line of a Stan .csv file with the inverse mass matrix
diagonal in it.
}
