% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_init_buffer}
\alias{get_init_buffer}
\title{init_buffer adaptation parameter}
\usage{
get_init_buffer(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
init_buffer adaptation parameter
}
