% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-helpers.R
\name{get_initial_number}
\alias{get_initial_number}
\title{Find number at the head of a string.}
\usage{
get_initial_number(s)
}
\arguments{
\item{s}{string}
}
\value{
number only.
}
\description{
Find number at the head of a string.
}
