% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_initial_stepsize}
\alias{get_initial_stepsize}
\title{initial stepsize parameter}
\usage{
get_initial_stepsize(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
initial stepsize parameter
}
