% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{get_job}
\alias{get_job}
\title{Retrieve a job with defaults integrated.}
\usage{
get_job(instructions, i, logger)
}
\arguments{
\item{instructions, }{loaded from .yaml file (list)}

\item{i}{pull out i'th job}

\item{logger}{used to log output}
}
\value{
job
}
\description{
Retrieve a job with defaults integrated.
}
