% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_kappa}
\alias{get_kappa}
\title{Kappa adaptation parameter}
\usage{
get_kappa(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
Kappa adaptation parameter
}
