% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_max_depth}
\alias{get_max_depth}
\title{max tree depth parameter}
\usage{
get_max_depth(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
max tree depth parameter
}
