% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_method_name}
\alias{get_method_name}
\title{Sampling method name from processed list.}
\usage{
get_method_name(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
method name
}
\description{
Sampling method name from processed list.
}
