% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_metric}
\alias{get_metric}
\title{metric used parameter.}
\usage{
get_metric(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
metric used parameter.
}
