% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_model_name}
\alias{get_model_name}
\title{Model name from list.}
\usage{
get_model_name(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
model name
}
\description{
Model name from list.
}
