% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_num_draws}
\alias{get_num_draws}
\title{Number of total iterations}
\usage{
get_num_draws(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
integer number of total iterations.
}
\description{
Number of total iterations
}
