% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_num_warmup}
\alias{get_num_warmup}
\title{Number of warmup iterations.}
\usage{
get_num_warmup(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
integer number of warmup iterations
}
\description{
Number of warmup iterations.
}
