% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_save_warmup}
\alias{get_save_warmup}
\title{Indicator whether to save warmup.}
\usage{
get_save_warmup(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
indicator as TRUE/FALSE
}
\description{
Indicator whether to save warmup.
}
