% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{get_script}
\alias{get_script}
\title{Retrieve the script to run on.}
\usage{
get_script(job, logger)
}
\arguments{
\item{job}{job to get script from}

\item{logger}{used to log output}
}
\value{
path to job script
}
\description{
Retrieve the script to run on.
}
