% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cn-helpers.R
\name{get_split_column_indexes}
\alias{get_split_column_indexes}
\title{Get list of column indexes of all parameters.}
\usage{
get_split_column_indexes(cn)
}
\arguments{
\item{cn}{Stan-format column names.}
}
\value{
list of named vectors, one entry per parameter, 
        each vector element is the index of a column
        for the parameter.
}
\description{
Get list of column indexes of all parameters.
}
