% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cn-helpers.R
\name{get_split_column_names}
\alias{get_split_column_names}
\title{Get list of column names of all parameters}
\usage{
get_split_column_names(cn)
}
\arguments{
\item{cn}{Stan-format column names.}
}
\value{
list of named vector, one entry per parameter,
        each vector element is the name of a column for
        that parameter
}
\description{
Get list of column names of all parameters
}
