% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_stan_version}
\alias{get_stan_version}
\title{Get the Stan version from processed list.}
\usage{
get_stan_version(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
stan version string.
}
\description{
Get the Stan version from processed list.
}
