% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_term_buffer}
\alias{get_term_buffer}
\title{terminal buffer adaptation parameter}
\usage{
get_term_buffer(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
terminal buffer adaptation parameter
}
