% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-helpers.R
\name{get_total_time}
\alias{get_total_time}
\title{Total run time}
\usage{
get_total_time(lines)
}
\arguments{
\item{CmdStan}{output as vector of lines}
}
\value{
total time to run model
}
\description{
Total run time
}
