% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cl-helpers.R
\name{get_window}
\alias{get_window}
\title{adaptation window parameter}
\usage{
get_window(cl)
}
\arguments{
\item{processed}{list from .csv}
}
\value{
parameter value
}
\description{
adaptation window parameter
}
