% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grab_comment_lines}
\alias{grab_comment_lines}
\title{Grab the lines of a Stan .csv file with the metadata}
\usage{
grab_comment_lines(lines)
}
\arguments{
\item{lines}{lines from the Stan .csv file, need only enough to get past the recorded meta-data.}
}
\value{
only the lines starting with a hash.
}
\description{
Grab the lines of a Stan .csv file with the metadata
}

