% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-compile.R
\name{hash_match}
\alias{hash_match}
\title{Check if two hashes match.}
\usage{
hash_match(a, b)
}
\arguments{
\item{a, }{an openssl::sha256 hash}

\item{b, }{an openssl::sha256 hash}
}
\value{
TRUE only if they are equal
}
\description{
Check if two hashes match.
}
