% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R
\name{index_hashes}
\alias{index_hashes}
\title{List all hashes under a target directory}
\usage{
index_hashes(target_dir)
}
\arguments{
\item{target_dir}{directory to search for hashes}
}
\value{
character vector of hashes
}
\description{
List all hashes under a target directory
}
