% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{insert_in}
\alias{insert_in}
\title{Add a key to a registry, with its arg-tree}
\usage{
insert_in(args, registry)
}
\arguments{
\item{args}{the art-tree to add}

\item{registry}{the registry to add to}
}
\value{
the registry with the added key
}
\description{
Add a key to a registry, with its arg-tree
}
