% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-install.R
\name{install_cmdstan}
\alias{install_cmdstan}
\title{Download CmdStan from github.com/stan-dev/cmdstan and build it in a
local directory.}
\usage{
install_cmdstan(target_dir = NULL)
}
\arguments{
\item{target_dir}{directory to install/build CmdStan in.}
}
\value{
path to .yaml file listing config details.
}
\description{
Download CmdStan from github.com/stan-dev/cmdstan and build it in a
local directory.
}
