% Generated by roxygen2 (4.0.2): do not edit by hand
\name{internal_parameters}
\alias{internal_parameters}
\title{Subset a vector of names to keep only sampler/internal parameters.}
\usage{
internal_parameters(x)
}
\arguments{
\item{x}{vector of names}
}
\value{
x, excluding names not ending in '__'.
}
\description{
Subset a vector of names to keep only sampler/internal parameters.
}

