% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{is_directory}
\alias{is_directory}
\title{Check if a path exists and is a directory}
\usage{
is_directory(p)
}
\arguments{
\item{p}{path to check}
}
\value{
TRUE iff conditions are met
}
\description{
Check if a path exists and is a directory
}
