% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{is_empty_list}
\alias{is_empty_list}
\title{Check whether list elements are all zero.}
\usage{
is_empty_list(x)
}
\arguments{
\item{x}{item to check}
}
\value{
TRUE only if x is a list.
}
\description{
Check whether list elements are all zero.
}
