% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{is_named_data_frame}
\alias{is_named_data_frame}
\title{Check if input is a data frame with column names.}
\usage{
is_named_data_frame(x)
}
\arguments{
\item{x}{input}
}
\value{
TRUE iff conditions are met
}
\description{
Check if input is a data frame with column names.
}
