% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{is_ragged_list}
\alias{is_ragged_list}
\title{Check whether list elements have different lenghts.}
\usage{
is_ragged_list(x)
}
\arguments{
\item{x}{item to check.}
}
\value{
TRUE only if x is a ragged list.
}
\description{
Check whether list elements have different lenghts.
}
