% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{label}
\alias{label}
\title{Create labelled samples.}
\usage{
label(A, labels = NULL)
}
\arguments{
\item{labels}{list with a named elements for each array margin
and each element is a data frame of groupings to join with
that margin's index.}

\item{array}{array of samples extracted from sample file set}
}
\value{
list with a component for a matrix of values (one group
        per row, one iteration per column), and a data.frame
        with string labels for each row.
}
\description{
Create labelled samples.
}
