% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{label_array}
\alias{label_array}
\title{Label parameter array}
\usage{
label_array(A, labels)
}
\arguments{
\item{labels, }{list, matching array dimensions, used to
label array margins, one list element per dimension, 
name is the dimension name, each entry is a vector
with labels or each entry of the dimension index.}

\item{array, }{Stan-dimension (+1 for iteration) array of 
parameter samples}
}
\description{
Label parameter array
}
