% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-trees.R
\name{load_args}
\alias{load_args}
\title{Load a Stannis run control list of arg-trees from 
a .yaml file}
\usage{
load_args(file)
}
\arguments{
\item{file}{a .yaml file with arg-tree structure
and defaults.}
}
\value{
list of arg-trees with hashes, not finalized.
}
\description{
Format description: the two top-level items are
'defaults', and 'runs'.  The each item under those
is a Stannis arg-tree with CmdStan arguments.
}
\details{
Arg-trees are loaded, each run arg-tree is merged with the
defaults, and a hash is added to each.
}
