% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{load_registry}
\alias{load_registry}
\title{Load a registry (or initialize if necessary).}
\usage{
load_registry(path = "registry.yaml")
}
\arguments{
\item{path}{path where to load the registry from}
}
\value{
the registry object
}
\description{
Load a registry (or initialize if necessary).
}
