% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-yaml.R
\name{load_yaml_args}
\alias{load_yaml_args}
\title{Load a list of arg-trees from a .yaml file}
\usage{
load_yaml_args(file)
}
\arguments{
\item{file}{a .yaml file with arg-tree structure
and defaults.}

\item{list}{of arg-trees with hashes, not finalized.}
}
\description{
Load a list of arg-trees from a .yaml file
}
