% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{logger}
\alias{logger}
\title{Return a closure with a log file set}
\usage{
logger(log_file = tempfile())
}
\arguments{
\item{log_file}{where to write log to}
}
\value{
closure, callable in various ways with strings.
}
\description{
Return a closure with a log file set
}
