% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-set-helpers.R
\name{merge_chains}
\alias{merge_chains}
\title{Merge chains}
\usage{
merge_chains(set)
}
\arguments{
\item{set}{object created by `read_stan_set`}
}
\value{
set object with only one merged chain.
}
\description{
Merge chains
}
