% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-yaml.R
\name{merge_lists}
\alias{merge_lists}
\title{Combine two lists by merging matching sub-trees.}
\usage{
merge_lists(...)
}
\arguments{
\item{...}{arbitrary number of lists.}
}
\value{
single merged list
}
\description{
Combine two lists by merging matching sub-trees.
}
