% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-trees.R
\name{merge_trees}
\alias{merge_trees}
\title{Combine multiple arg-trees by merging matching sub-trees.}
\usage{
merge_trees(...)
}
\arguments{
\item{...}{arbitrary number of lists.}
}
\value{
single merged list
}
\description{
Combine multiple arg-trees by merging matching sub-trees.
}
