% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{model_parameters}
\alias{model_parameters}
\title{Subset a vector of names to keep only model parameters.}
\usage{
model_parameters(s)
}
\arguments{
\item{s}{vector of names}
}
\value{
s, excluding names ending in '__'.
}
\description{
Subset a vector of names to keep only model parameters.
}
