% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor runs in a given root directory}
\usage{
monitor(root = ".", split_complete = TRUE, ...)
}
\arguments{
\item{root}{directory to search for runs under.}

\item{...}{other arguments passed to find_run_files}
}
\value{
data.frame with hash, path, and file names of output files
        with S3 class c("mcmc.monitor", "data.frame" to simplify
        printing.
}
\description{
Monitor runs in a given root directory
}
