% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-helpers.R
\name{name_margins}
\alias{name_margins}
\title{Create margin labels following a pattern}
\usage{
name_margins(A)
}
\arguments{
\item{A}{array to create labels for (or matrix).}
}
\value{
list of margin labels
}
\description{
Create margin labels following a pattern
}
