% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-helpers.R
\name{normalize_margin_names}
\alias{normalize_margin_names}
\title{Normalize an array's labels.}
\usage{
normalize_margin_names(A)
}
\arguments{
\item{A}{array to create labels for (or matrix).}
}
\value{
array with consitent dimnames.
}
\description{
Normalize an array's labels.
}
