% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checking.R
\name{not_null}
\alias{not_null}
\title{Return an NA or the value}
\usage{
not_null(x)
}
\arguments{
\item{x}{a list element, could be NULL}
}
\value{
NA or the item
}
\description{
Return an NA or the value
}
