% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-helpers.R
\name{padded_sequence}
\alias{padded_sequence}
\title{Take a numeric sequence and force it to sort properly as a character
sequence by padding to length.}
\usage{
padded_sequence(x)
}
\arguments{
\item{x}{numeric sequence}
}
\value{
sequence of strings padded out to the same length
}
\description{
Take a numeric sequence and force it to sort properly as a character
sequence by padding to length.
}
