% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_parameters}
\alias{parse_parameters}
\title{Takes names from Stan .csv files and produces an object with parsed
info from comments as well as parameter/transformed parameter names,
a matched R object class, and object dimensions.}
\usage{
parse_parameters(file = NULL)
}
\arguments{
\item{file}{A vector of length 1, containing the file name of a Stan .csv output file.}
}
\value{
A list with metadata on the Stan run recorded in the file.
}
\description{
Takes names from Stan .csv files and produces an object with parsed
info from comments as well as parameter/transformed parameter names,
a matched R object class, and object dimensions.
}
\examples{
parameters <- parse_parameters(file='/tmp/stan.csv')
}

