% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-helpers.R
\name{path_split}
\alias{path_split}
\title{Normalizes a path and splits it on a path separator, 
typically '/'.}
\usage{
path_split(path, separator = "/")
}
\arguments{
\item{path}{non-normalized paths.}

\item{separator}{the path separator, typically '/'}
}
\value{
list of normalized and split path parts.
}
\description{
Normalizes a path and splits it on a path separator, 
typically '/'.
}
