% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-trees.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Prepare a single input tree by combining paths.}
\usage{
prepare_inputs(args)
}
\arguments{
\item{args}{an arg-tree}
}
\value{
an arg-tree
}
\description{
Prepare a single input tree by combining paths.
}
