% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{print.mcmc.monitor}
\alias{print.mcmc.monitor}
\title{Print output of stannis::monitor}
\usage{
\method{print}{mcmc.monitor}(completion)
}
\arguments{
\item{completion}{data from 'monitor' on the completion of runs.}
}
\description{
Print output of stannis::monitor
}
