% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_bash_array}
\alias{process_bash_array}
\title{Reads a simple "a" = "b" format, ignoring non-matching data.}
\usage{
process_bash_array(file = "model-data.sh")
}
\arguments{
\item{file}{source file.}
}
\value{
named list with 'a' for names and 'b' for values.
}
\description{
Reads a simple "a" = "b" format, ignoring non-matching data.
}

