% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sets.R
\name{process_diagnostics}
\alias{process_diagnostics}
\title{Process sampling file
Process diagnostics}
\usage{
process_diagnostics(x)
}
\arguments{
\item{x}{`read_stan_csv` output of diagnostic .csv file.}
}
\description{
Process sampling file
Process diagnostics
}
