% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{process_key_value}
\alias{process_key_value}
\title{Processes a strings data format (key[split]value)}
\usage{
process_key_value(x, split = "=")
}
\arguments{
\item{x}{vector of strings to process}

\item{split}{string to distinguish lhs from rhs}
}
\value{
named vector with values as values and tags as names.
}
\description{
Processes a strings data format (key[split]value)
}
