% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_adapt}
\alias{push_adapt}
\title{Add sampler adaptation sub-arguments.}
\usage{
push_adapt(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add sampler adaptation sub-arguments.
}
