% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_algorithm}
\alias{push_algorithm}
\title{Add algorithm sub-arguments.}
\usage{
push_algorithm(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add algorithm sub-arguments.
}
