% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_arg}
\alias{push_arg}
\title{Basic method to return one CmdStan argument.}
\usage{
push_arg(name, value)
}
\arguments{
\item{name}{name of the parameter to add.}

\item{value}{value the parameter takes.}
}
\value{
combined string argument.
}
\description{
Basic method to return one CmdStan argument.
}
