% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_engine}
\alias{push_engine}
\title{Add HMC engine sub-arguments}
\usage{
push_engine(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add HMC engine sub-arguments
}
