% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_fixed}
\alias{push_fixed}
\title{Add fixed_param non-sampler arguments.}
\usage{
push_fixed(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add fixed_param non-sampler arguments.
}
