% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_hmc}
\alias{push_hmc}
\title{Add HMC algorithm sub-arguments}
\usage{
push_hmc(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add HMC algorithm sub-arguments
}
