% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_metric}
\alias{push_metric}
\title{Add HMC metric sub-arguments.}
\usage{
push_metric(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add HMC metric sub-arguments.
}
