% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_nuts}
\alias{push_nuts}
\title{add nuts sampler sub-arguments.}
\usage{
push_nuts(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
add nuts sampler sub-arguments.
}
