% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_random}
\alias{push_random}
\title{Add a random init seed argument}
\usage{
push_random(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add a random init seed argument
}
