% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_sample}
\alias{push_sample}
\title{Add sampler sub-arguments}
\usage{
push_sample(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
Add sampler sub-arguments
}
