% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_simple}
\alias{push_simple}
\title{Basic method to pull out arguments by name
from a list.}
\usage{
push_simple(args, names)
}
\arguments{
\item{args}{(potentially nested) list containing arguments.}

\item{names}{vector of named arguments to pull out.}
}
\value{
turn vector of arguments as strings
}
\description{
Basic method to pull out arguments by name
from a list.
}
