% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-args.R
\name{push_static}
\alias{push_static}
\title{add static HMC sub-arguments.}
\usage{
push_static(args)
}
\arguments{
\item{args}{argument list}
}
\value{
argument string
}
\description{
add static HMC sub-arguments.
}
