% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sets.R
\name{read_file_set}
\alias{read_file_set}
\title{Read a set of Stan files and their metadata}
\usage{
read_file_set(root = ".", hashes = NULL, control = "finalized.yaml",
  samples = "output.csv", diagnostics = "diagnostics.csv", ...)
}
\arguments{
\item{search}{directories to read from}

\item{pattern}{pattern of filenames to read}
}
\value{
a processed and merged list of files.
}
\description{
Read a set of Stan files and their metadata
}
