% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-helpers.R
\name{read_key_value}
\alias{read_key_value}
\title{Reads a simple tag:value file format.}
\usage{
read_key_value(file, split = "=")
}
\arguments{
\item{file}{read all lines from here}

\item{split}{string to distinguish lhs from rhs}
}
\value{
named vector with values as values and tags as names.
}
\description{
Reads a simple tag:value file format.
}
