% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{read_record_table}
\alias{read_record_table}
\title{Read the record table:}
\usage{
read_record_table(file)
}
\arguments{
\item{file}{file holding record table.}
}
\value{
record table as data.frame
}
\description{
Read the record table:
}
