% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-run.R
\name{read_run}
\alias{read_run}
\title{Read a single Stannis run, rewrite to binary format, and return 
dimensions and paths to outptu.}
\usage{
read_run(root = ".", hash = NULL, uuid = NULL,
  control_file = "finalized.yaml", sample_file = "output.csv",
  diagnostic_file = "diagnostics.csv")
}
\arguments{
\item{root}{where the run directories are rooted}

\item{hash}{hash for this run.}

\item{uuid}{UUID for this run.}

\item{control_file}{what the .yaml with run settings is called}

\item{sample_file}{what the sample output file is called}

\item{diagnostic_file}{file name for diagnostic output}
}
\value{
list with run description
}
\description{
Read a single Stannis run, rewrite to binary format, and return 
dimensions and paths to outptu.
}
