% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-file-readers.R
\name{read_stan}
\alias{read_stan}
\title{Read Stan file and metadata}
\usage{
read_stan(file)
}
\arguments{
\item{file}{path}
}
\value{
list with samples and metadata
}
\description{
Read Stan file and metadata
}
