% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-file-readers.R
\name{read_stan_metadata}
\alias{read_stan_metadata}
\title{Properly read a stan file, avoiding factors and comments, including
reading column names from a header.}
\usage{
read_stan_metadata(file)
}
\arguments{
\item{file}{An output file produced by CmdStan.}
}
\value{
A data.frame.
}
\description{
Properly read a stan file, avoiding factors and comments, including
reading column names from a header.
}
