% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{record_table}
\alias{record_table}
\title{Record basic info on a run that indexes the registry}
\usage{
record_table(args)
}
\arguments{
\item{args}{arg-tree object (list)}
}
\value{
data.frame of the record table
}
\description{
Record basic info on a run that indexes the registry
}
