% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_run}
\alias{register_run}
\title{Record that a run is being started}
\usage{
register_run(args)
}
\arguments{
\item{args}{arg-tree object (list)}
}
\description{
Record that a run is being started
}
