% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{remove_failed_runs}
\alias{remove_failed_runs}
\title{Find and remove run directories with no output file.}
\usage{
remove_failed_runs(root, ...)
}
\arguments{
\item{root}{root path to search under}
}
\value{
hashes of removed directories
}
\description{
Find and remove run directories with no output file.
}
