% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{remove_key}
\alias{remove_key}
\title{Remove a record from the registry}
\usage{
remove_key(registry, hash)
}
\arguments{
\item{registry}{a registry object}

\item{hash}{the key (hash) to remove}
}
\description{
Remove a record from the registry
}
