% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-arg-tree-helpers.R
\name{replicate_args}
\alias{replicate_args}
\title{Replicate arg-tree and add replicate tag.}
\usage{
replicate_args(args)
}
\arguments{
\item{args}{agr-tree object (list).}
}
\value{
list of arg-trees, separated s.t. each arg-tree
        is for one replicate
}
\description{
Replicate arg-tree and add replicate tag.
}
