% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-rewrite.R
\name{rewrite_stan_csv}
\alias{rewrite_stan_csv}
\title{Rewrite a Stan .csv output file.}
\usage{
rewrite_stan_csv(source, root, tag, comment)
}
\arguments{
\item{source}{path to Stan format .csv file}

\item{root}{where to rewrite to}

\item{tag}{uuid that gets embedded in binary output}

\item{comment}{freeform text embedded in binary output}
}
\value{
TRUE if fully rewritten
}
\description{
Rewrite a Stan .csv output file.
}
