% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scripts.R
\name{run_data_shim}
\alias{run_data_shim}
\title{Run a simple R script based on one input and one set of transformations.}
\usage{
run_data_shim(run)
}
\arguments{
\item{run}{a list with elements for the file listing transformations and
requirements as well as where to find them.}
}
\value{
TRUE if it completes... output goes into a file...
}
\description{
Run a simple R script based on one input and one set of transformations.
}
