% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scripts.R
\name{run_isolated_script}
\alias{run_isolated_script}
\title{Run a script in an isolated environment and move outputs into a target
directory}
\usage{
run_isolated_script(run, e = new.env(parent = parent.env(.GlobalEnv)))
}
\arguments{
\item{run}{list describing the job}

\item{e}{environment to isolate in.}
}
\value{
e, with cruft from script.
}
\description{
Run a script in an isolated environment and move outputs into a target
directory
}
