% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-run.R
\name{run_model_cmd}
\alias{run_model_cmd}
\title{Run a CmdStan model based on a single arg-tree.}
\usage{
run_model_cmd(...)
}
\arguments{
\item{...}{sub-tree of arg-tree split into arguments.}
}
\value{
NULL, run as a system command.
}
\description{
Run a CmdStan model based on a single arg-tree.
}
