% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scripts.R
\name{run_script_yaml}
\alias{run_script_yaml}
\title{Run either a shell script, a simple data-shim or a full
isolated script.  This is all too complicated. :/}
\usage{
run_script_yaml(file)
}
\arguments{
\item{.yaml}{file with control data describing inputs
outputs and targets along with the script to run.}
}
\description{
Run either a shell script, a simple data-shim or a full
isolated script.  This is all too complicated. :/
}
