% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-run.R
\name{run_yaml}
\alias{run_yaml}
\title{Run a set of CmdStan runs based on a set of arg-trees
in a .yaml file.}
\usage{
run_yaml(file, cores = getOption("cl.cores", 1))
}
\arguments{
\item{file}{.yaml file with instructions}

\item{cores}{number of cores to use for the run}
}
\description{
Run a set of CmdStan runs based on a set of arg-trees
in a .yaml file.
}
