% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{save_output}
\alias{save_output}
\title{Save output}
\usage{
save_output(job, output, logger)
}
\arguments{
\item{job}{job description}

\item{output}{list of output objects}

\item{logger}{logger to write log to...}
}
\description{
Save output
}
