% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{save_registry}
\alias{save_registry}
\title{Save a registry}
\usage{
save_registry(registry, path = "registry.yaml")
}
\arguments{
\item{registry}{a registry object}

\item{path}{the path to save the registry to (.yaml)}
}
\value{
path where the reigstry was saved
}
\description{
Save a registry
}
