% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sets.R
\name{scatter}
\alias{scatter}
\title{Scatter set}
\usage{
scatter(set, target)
}
\arguments{
\item{set}{set to write out.}

\item{target}{target directory to write to.}
}
\value{
target directory
}
\description{
Scatter set
}
