% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{scripted}
\alias{scripted}
\title{Run files based on scripts... take N+1...}
\usage{
scripted(file, log_file = tempfile(), debug = FALSE)
}
\arguments{
\item{file}{.yaml file with instructions, see example}

\item{log_file}{where to write text log to.}
}
\value{
log_file where logs were written.
}
\description{
Run files based on scripts... take N+1...
}
