% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-examples.R
\name{setup_runs}
\alias{setup_runs}
\title{Set up a set of file to run examples.}
\usage{
setup_runs(target_dir)
}
\arguments{
\item{target_dir}{directory to set up in.}
}
\value{
TRUE iff successful
}
\description{
Set up a set of file to run examples.
}
