% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-df-helpers.R
\name{split_data_by_parameter}
\alias{split_data_by_parameter}
\title{Split Stan-style data from a .csv file by parameter.}
\usage{
split_data_by_parameter(data)
}
\arguments{
\item{x}{Stan-format .csv parameter data.}
}
\value{
list of Stan-format parameter data, with one 
        entry (data.frame) per parameter
}
\description{
Split Stan-style data from a .csv file by parameter.
}
