\name{stannis-package}
\alias{stannis-package}
\alias{stannis}
\docType{package}
\title{
\packageTitle{stannis}
}
\description{
\packageDescription{stannis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stannis}
\packageIndices{stannis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stannis}

Maintainer: \packageMaintainer{stannis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
