% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stannis-package.R, R/stannis.R
\docType{package}
\name{stannis}
\alias{stannis}
\alias{stannis-package}
\alias{stannis}
\title{Bayesian Inference of State Space Models}
\usage{
stannis(formula, data, level, slope, seasonal, a1, P1, beta, iter, thin = 1,
  nsim_states = 10, stan_inits, n_threads = 1, simplify_output = TRUE,
  is_seed = sample.int(.Machine$integer.max, 1), ...)
}
\arguments{
\item{formula}{Object of class \code{formula} which defines the regression part. 
Or a single vector of observations in case of no covariates.}

\item{data}{Optional data frame (or object coercible to such) containing the data.}

\item{level}{A vector of length two, defining the mean and standard deviation of the 
truncated Gaussian prior for the standard deviation of the level noise term.}

\item{slope}{A vector of length two, defining the mean and standard deviation of the 
truncated Gaussian prior for the standard deviation of the slope noise term.}

\item{seasonal}{A vector of length two, defining the mean and standard deviation of the 
truncated Gaussian prior for the standard deviation of the seasonal noise term.}

\item{a1}{Prior mean for the state vector at time 1.}

\item{P1}{Prior covariance matrix for state vector at time 1.}

\item{beta}{A matrix with \eqn{k} rows and 2 columns, where first columns defines the 
prior means of the Gaussian priors of the corresponding \eqn{k} regression coefficients, 
and the second column defines the the standard deviations of those prior distributions.}

\item{iter}{Number of iterations for the MCMC.}

\item{thin}{Thinning interval used in MCMC.}

\item{nsim_states}{Number of particles used in particle filter.}

\item{stan_inits}{Initial values for the MCMC as a list of lists.}

\item{n_threads}{Number of threads to use in parallel.}

\item{simplify_output}{If \code{TRUE} (default), does not return the original output from \code{Stan}.}

\item{is_seed}{Seed for the importance sampling correction.}
}
\description{
Bayesian inference of exponential family state space models
where the state dynamics are Gaussian but the observational density is either
Gaussian, Poisson, binomial, or negative binomial. The novel approach used by
the stannis package is based on the combination of fast Gaussian approximation,
efficient No-U-Turn sampler provided by Stan, and the parallelisable importance
sampling type corrected Markov chain Monte Carlo approach.

\code{stannis} performs fully Bayesian inference of structural Poisson time series optionally with covariates.
}
\details{
\code{stannis} first obtains approximate marginal posterior using Markov chain Monte Carlo via 
Hamiltonian Monte Carlo sampler of \code{Stan}, and then performs importance sampling type correction of this 
posterior sample in order to obtain "exact" joint posterior. The correction step is done using particle filtering.
}
