% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-helpers.R
\name{std_estimates}
\alias{std_estimates}
\title{Standard summary function.}
\usage{
std_estimates(x)
}
\arguments{
\item{x}{per-iteration parameter values.}
}
\value{
per-estimate summary.
}
\description{
Standard summary function.
}
