% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{substitutions}
\alias{substitutions}
\title{Substitute tags for the referenced file content.}
\usage{
substitutions(model = NULL, search = NULL, output = NULL)
}
\arguments{
\item{model}{full path to model file with tags in it.  Must be specified.}

\item{search}{path to search for chunks to be inserted when tags
are replaced. If not specified search is restricted to model
directory.}

\item{output}{full path to file that model will be 
written into. If not specified it is derived from the
model path by replacing the extension with .stan}
}
\description{
Substitute tags for the referenced file content.
}
