% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{tag_file}
\alias{tag_file}
\title{File matching a tag.  Replaces '::' separator with '-' separator.}
\usage{
tag_file(tag)
}
\arguments{
\item{tag}{text descriptor of file name.}
}
\value{
implied file name
}
\description{
File matching a tag.  Replaces '::' separator with '-' separator.
}
