% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{tag_find}
\alias{tag_find}
\title{Find which items of a file read in as a character vector
include tags.}
\usage{
tag_find(text)
}
\arguments{
\item{text}{file as a character vector (1 element per line).}
}
\value{
vector of line numbers with tags.
}
\description{
Find which items of a file read in as a character vector
include tags.
}
