% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice-model.R
\name{tag_pull}
\alias{tag_pull}
\title{Extract the tag from file read in as a vector of text strings.}
\usage{
tag_pull(text)
}
\arguments{
\item{text}{file as a vector of text strings.}
}
\value{
all tags
}
\description{
Extract the tag from file read in as a vector of text strings.
}
