% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{trim_comment_char}
\alias{trim_comment_char}
\title{Trims comment char from the front of a string.}
\usage{
trim_comment_char(s, c = "#")
}
\arguments{
\item{s}{the string}

\item{c}{comment character}
}
\value{
s without leading and trailing whitespace.
}
\description{
Trims comment char from the front of a string.
}
