% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-set-helpers.R
\name{trim_warmup}
\alias{trim_warmup}
\title{Trim warmup}
\usage{
trim_warmup(set)
}
\arguments{
\item{set}{object created by `read_stan_set`}
}
\value{
same object but with warmup iterations removed.
}
\description{
Trim warmup
}
