% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{unroll_array}
\alias{unroll_array}
\title{Array to long-form}
\usage{
unroll_array(A, labels = NULL)
}
\arguments{
\item{labels}{list with a named elements for each array margin
and each element lists the names for that margin.}

\item{array}{array of samples extracted from sample file set}
}
\value{
list with a component for a matrix of values (one group
        per row, one iteration per column), and a data.frame
        with string labels for each row.
}
\description{
Array to long-form
}
