% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{unroll_array_to_df}
\alias{unroll_array_to_df}
\title{Array to long-form, combined into one data frame.}
\usage{
unroll_array_to_df(A, labels = NULL)
}
\arguments{
\item{labels}{list with a named elements for each array margin
and each element lists the names for that margin.}

\item{array}{array of samples extracted from sample file set}
}
\value{
data frame with grouping variables and the associated
        values (iterations are in columns with numerical names.
}
\description{
Array to long-form, combined into one data frame.
}
