% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-helpers.R
\name{uuid}
\alias{uuid}
\title{Turn any vector of strings into a UUID}
\usage{
uuid(s = NULL)
}
\arguments{
\item{s}{vector of strings}
}
\value{
UUID
}
\description{
Turn any vector of strings into a UUID
}
