% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-r-hat.R
\name{worst_psrf}
\alias{worst_psrf}
\title{Get the worst R-hat values per parameter.}
\usage{
worst_psrf(x)
}
\arguments{
\item{list}{of r-hat values, per-parameter}
}
\value{
tail of sorted list of r-hat values, per-parameter
}
\description{
Get the worst R-hat values per parameter.
}
