% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{write_record_table}
\alias{write_record_table}
\title{Write the record table:}
\usage{
write_record_table(x, file)
}
\arguments{
\item{x}{record table as data.frame}

\item{file}{to write record table to}
}
\description{
Write the record table:
}
