% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R
\name{create_hash}
\alias{create_hash}
\title{create a pseudo-'secret', salted with project_id}
\usage{
create_hash(args)
}
\arguments{
\item{args}{arg-tree object (list).}
}
\value{
hash
}
\description{
create a pseudo-'secret', salted with project_id
}
