% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{find_file}
\alias{find_file}
\title{Find a full path to a file name, recursively}
\usage{
find_file(root, name, complete = TRUE)
}
\arguments{
\item{root}{search everything below this directory}

\item{name}{file name to search for}
}
\value{
full normalized path to file
}
\description{
Find a full path to a file name, recursively
}
