% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-cn-helpers.R
\name{generate_dims}
\alias{generate_dims}
\title{Create a dimlist out of a character vector of Stan-style column
names.}
\usage{
generate_dims(x)
}
\arguments{
\item{x}{vector of column names for a given parameter.}
}
\value{
an un-named vector of dimensions.
}
\description{
Create a dimlist out of a character vector of Stan-style column
names.
}
