% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-read.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Read the parameter dimensions of the binary rewrite}
\usage{
get_dimensions(root, name = NULL)
}
\arguments{
\item{root}{where the rewrite is rooted.}

\item{name}{optionally name of a parameter to get dimensions from.}
}
\value{
named list of parameter dimensions
}
\description{
Read the parameter dimensions of the binary rewrite
}
