% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_parameter}
\alias{get_parameter}
\title{Read a specific parameter from the binary rewrite}
\usage{
get_parameter(root, name)
}
\arguments{
\item{root}{where the rewrite is rooted}
}
\value{
array of parameter samples
}
\description{
Read a specific parameter from the binary rewrite
}
