% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sets.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Get names of parameters in set}
\usage{
get_parameter_names(set)
}
\arguments{
\item{set}{object created by read_file_set}
}
\value{
names of parameters in set
}
\description{
Get names of parameters in set
}
