% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{internal_parameters}
\alias{internal_parameters}
\title{Subset a vector of names to keep only sampler/internal parameters.}
\usage{
internal_parameters(s)
}
\arguments{
\item{s}{vector of names}
}
\value{
s, including names ending in '__'.
}
\description{
Subset a vector of names to keep only sampler/internal parameters.
}
