% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model_parameters}
\alias{model_parameters}
\title{Subset a vector of names to keep only model parameters.}
\usage{
model_parameters(x)
}
\arguments{
\item{x}{vector of names}
}
\value{
x, excluding names ending in '__'.
}
\description{
Subset a vector of names to keep only model parameters.
}

