% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-set-helpers.R
\name{read_file_set}
\alias{read_file_set}
\title{Read a set of Stan files and their metadata}
\usage{
read_file_set(root = ".", pattern)
}
\arguments{
\item{root}{directory to read from}

\item{pattern}{pattern of filenames to read}
}
\value{
a processed and merged list of files.
}
\description{
Read a set of Stan files and their metadata
}
