% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-set-helpers.R
\name{read_stan_csv}
\alias{read_stan_csv}
\title{Wrapper for C++ function wrapper.}
\usage{
read_stan_csv(file)
}
\arguments{
\item{file}{file name passed to C++ reader.}
}
\value{
list with samples, indexing, etc...
}
\description{
Wrapper for C++ function wrapper.
}
