% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stannis-package.R
\docType{package}
\name{stannis}
\alias{stannis}
\alias{stannis-package}
\title{Bayesian Inference of State Space Models}
\description{
Bayesian inference of exponential family state space models
where the state dynamics are Gaussian but the observational density is either
Gaussian, Poisson, binomial, or negative binomial. The novel approach used by
the stannis package is based on the combination of fast Gaussian approximation,
efficient No-U-Turn sampler provided by Stan, and the parallelisable importance
sampling type corrected Markov chain Monte Carlo approach.
}
