% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-array-helpers.R
\name{summarize}
\alias{summarize}
\title{Summarize iterations to estimates!}
\usage{
summarize(x, f, ...)
}
\arguments{
\item{x}{list from with named elements 'values' and 'grouping'.  The 'values' 
element is a matrix (n-groups by n-iterations) with sampled
parameter values.  The 'grouping' element is a data.frame with 
n-groups rows and one column per index.}

\item{f}{function to use to aggregate (applied row-wise)}
}
\value{
same list but with iteration columns summarized to estimates.
}
\description{
Summarize iterations to estimates!
}
